fix:完善.gitignore文件，提交版本库时过滤掉某些不必要文件；修改文章浏览相关代码

+++ src/main/java/com/pxl/service/ArticleService.java
@@ -18,6 +18,6 @@ public interface ArticleService extends IService<Article> {
 
     List<ArticleVo> findAllByKeywords(String keywords);
 
-    Integer addViewNums(Integer id);
+    Integer addViewNums(Long id);
 
 }
